resolves #47 stop requiring specific include file scheme
support conversion of article documents to a single-chapter file
resolves #205 support special chapters like bibliography
resolves #190 `basedir` now points to spine document directory when processing chapter files
resolves #178 fix image and listing numbers being reset in each chapter
resolves #166 fix xref resolving between sub-includes of chapter files
resolves #151 add support for contentless include files
resolves #136 drop nonstandard `<<chapter#>>` xref syntax and instead support vanilla `<<anchor>>` or `<<file#anchor>>` syntax
resolves #206 properly include bibliography generated by asciidoctor-bibtex
require Asciidoctor 1.5.6+
