+++ .github/workflows/ci.yml
@@ -21,7 +21,7 @@ jobs:
       matrix:
         ruby: [jruby, 2.3, 2.4, 2.5, 2.6, 2.7]
         platform: [ubuntu-latest, macos-latest, windows-latest]
-        asciidoctor: [1.5.3, '']
+        asciidoctor: [1.5.6, '']
         exclude:
           # kindlegen fails to install on 2.3 on Windows. See https://github.com/asciidoctor/asciidoctor-epub3/pull/213
           - ruby: 2.3
