+++ CHANGELOG.adoc
@@ -7,6 +7,16 @@ For a detailed view of what has changed, refer to the {uri-repo}/commits/master[
 
 == Unreleased
 
+* support conversion of article documents to a single-chapter file
+* stop requiring specific include file scheme (#47)
+* support special chapters like bibliography (#205)
+* `basedir` now points to main document directory when processing chapter files (#190)
+* fix image and listing numbers being reset in each chapter (#178)
+* fix xref resolving between sub-includes of chapter files (#166)
+* add support for contentless include files (#151)
+* drop nonstandard `<<chapter#>>` xref syntax and instead support vanilla `<<anchor>>` or `<<file#anchor>>` syntax (#136)
+* properly include bibliography generated by asciidoctor-bibtex (#206)
+* require Asciidoctor 1.5.6+
 * do not include current date if `:reproducible:` attribute is set (#203)
 * respect `SOURCE_DATE_EPOCH` environment variable for reproducible builds
 * fix invalid markup produced for tables with footer (#295)
