+++ lib/asciidoctor-epub3.rb
@@ -5,7 +5,6 @@ require 'asciidoctor/extensions'
 require 'gepub'
 require_relative 'asciidoctor-epub3/ext'
 require_relative 'asciidoctor-epub3/converter'
-require_relative 'asciidoctor-epub3/packager'
 
 # We need to be able to write files with unicode names. See https://github.com/asciidoctor/asciidoctor-epub3/issues/217
 ::Zip.unicode_names = true
