+++ spec/fixtures/bibliography-chapter/bibliography.adoc
@@ -1,0 +1,7 @@
+[bibliography]
+= Bibliography
+
+- [[[pp]]] Andy Hunt & Dave Thomas. The Pragmatic Programmer:
+From Journeyman to Master. Addison-Wesley. 1999.
+- [[[gof,2]]] Erich Gamma, Richard Helm, Ralph Johnson & John Vlissides. Design Patterns:
+Elements of Reusable Object-Oriented Software. Addison-Wesley. 1994.
