+++ spec/fixtures/bibliography-chapter/chapter.adoc
@@ -1,0 +1,4 @@
+= Chapter
+
+_The Pragmatic Programmer_ <<pp>> should be required reading for all developers.
+To learn all about design patterns, refer to the book by the "`Gang of Four`" <<gof>>.
