+++ spec/fixtures/diagram/book.adoc
@@ -1,7 +1,6 @@
 = Diagrams book
 :doctype: book
 
-// TODO: https://github.com/asciidoctor/asciidoctor-epub3/issues/190
-// :plantumlconfig: ./plantuml.cfg
+:plantumlconfig: ./plantuml.cfg
 
 include::subdir/chapter.adoc[leveloffset=+1]
