+++ spec/fixtures/inter-chapter-xref/chapter-a.adoc
@@ -3,4 +3,4 @@
 This is chapter A.
 There's not much too it.
 
-Time to move on to <<chapter-b#>>.
+Time to move on to <<chapter-b>>.
