+++ spec/fixtures/inter-subchapter-xref/book.adoc
@@ -1,0 +1,8 @@
+= Book Title
+:doctype: book
+:idprefix:
+:idseparator: -
+
+include::chapter-a.adoc[leveloffset=+1]
+
+include::chapter-b.adoc[leveloffset=+1]
