+++ spec/fixtures/inter-subchapter-xref/chapter-a.adoc
@@ -1,0 +1,3 @@
+= Chapter A
+
+include::subchapter-a.adoc[leveloffset=+1]
