+++ spec/fixtures/inter-subchapter-xref/chapter-b.adoc
@@ -1,0 +1,3 @@
+= Chapter B
+
+include::subchapter-b.adoc[leveloffset=+1]
