+++ spec/fixtures/inter-subchapter-xref/subchapter-a.adoc
@@ -1,0 +1,3 @@
+= Subchapter A
+
+Look at <<subchapter-b#anchor>>.
