+++ spec/fixtures/inter-subchapter-xref/subchapter-b.adoc
@@ -1,0 +1,3 @@
+= Subchapter B
+
+[[anchor,label]]Anchors rock!
