+++ spec/fixtures/invalid.adoc
@@ -1,0 +1,3 @@
+= Title
+
+<<nonexistent,xref to nowhere>>
