+++ spec/fixtures/listing-chapter/book.adoc
@@ -1,0 +1,8 @@
+= Listing chapter book
+:doctype: book
+:idprefix:
+:idseparator: -
+
+include::chapter-a.adoc[leveloffset=+1]
+
+include::chapter-b.adoc[leveloffset=+1]
