+++ spec/fixtures/listing-chapter/chapter-a.adoc
@@ -1,0 +1,6 @@
+= Chapter A
+
+[title=.gitattributes]
+----
+*   -text
+----
