+++ spec/fixtures/listing-chapter/chapter-b.adoc
@@ -1,0 +1,6 @@
+= Chapter B
+
+[title=.gitattributes]
+----
+*   -text
+----
