+++ spec/fixtures/minimal/chapter.adoc
@@ -1,0 +1,1 @@
+= Chapter
